// admin/src/app/dashboard/course/types.ts
export enum CourseStatus {
  DRAFT = 'draft',
  PUBLISHED = 'published',
  ARCHIVED = 'archived'
}

export enum LessonType {
  THEORY = 'theory',
  TEST = 'test',
  PRACTICE = 'practice'
}

export enum MediaType {
  IMAGE = 'image',
  VIDEO = 'video',
  DOCUMENT = 'document'
}

// Course Types
export interface Course {
  id: number;
  title: string;
  description: string | null;
  author_id: number | null;
  status: CourseStatus;
  price: number;
  created_at: string;
  updated_at: string | null;
}